requirements.txt
pandas
ollama
openai
python-dotenv
google-generativeai
openai-agents
numpy
llama-index
cohere
streamlit
google-generativeai
python-dotenv